Issue 198: fix saving of bookmarks, Issue 204 Category JPA stuff moved to WeblogManager.
