+++ app/src/main/java/org/apache/roller/weblogger/ui/struts2/editor/EntryBean.java
@@ -25,7 +25,7 @@ import org.apache.commons.lang3.StringUtils;
 import org.apache.commons.logging.Log;
 import org.apache.commons.logging.LogFactory;
 import org.apache.roller.weblogger.WebloggerException;
-import org.apache.roller.weblogger.business.WeblogEntryManager;
+import org.apache.roller.weblogger.business.WeblogManager;
 import org.apache.roller.weblogger.business.WebloggerFactory;
 import org.apache.roller.weblogger.pojos.WeblogCategory;
 import org.apache.roller.weblogger.pojos.WeblogEntry;
@@ -311,7 +311,7 @@ public class EntryBean {
         if (getCategoryId() != null) {
             WeblogCategory cat = null;
             try {
-                WeblogEntryManager wmgr = WebloggerFactory.getWeblogger().getWeblogEntryManager();
+                WeblogManager wmgr = WebloggerFactory.getWeblogger().getWeblogManager();
                 cat = wmgr.getWeblogCategory(getCategoryId());
             } catch (WebloggerException ex) {
                 log.error("Error getting category by id", ex);
